loss upd
